+++ ios/Classes/ZegoLiveRoomPlugin.h
@@ -1,5 +1,6 @@
 #import <Flutter/Flutter.h>
 #import <ZegoLiveRoom/ZegoLiveRoom.h>
+#import <ZegoLiveRoom/ZegoLiveRoomApi-Publisher2.h>
 #import <ZegoLiveRoom/ZegoLiveRoomApi-AudioIO.h>
 #import <ZegoLiveRoom/zego-api-media-side-info-oc.h>
 #import <ZegoLiveRoom/zego-api-external-video-render-oc.h>